changed to hybrid proximity monitoring
