+++ app/src/main/java/com/beakoninc/locusnotes/di/AppModule.kt
@@ -86,9 +86,10 @@ object AppModule {
     fun provideProximityManager(
         @ApplicationContext context: Context,
         locationService: LocationService,
-        noteRepository: NoteRepository
+        noteRepository: NoteRepository,
+        activityRecognitionManager: ActivityRecognitionManager
     ): ProximityManager {
-        return ProximityManager(context, locationService, noteRepository)
+        return ProximityManager(context, locationService, noteRepository, activityRecognitionManager)
     }
 
 }
